Final SQL Stored Procedures and Tables
Here are all the stored procedures. Keep in mind that not all of these
may be in use by the application, and that ones not being used may not
be functional.

Also, there is one sql file in the 'Full_Database_Export folder, which
can be run to generate the entire database from scratch. This contains
all the create table statements in correct order.